# docs/swagger/openapi.yaml
openapi: "3.0.3"
info:
  title: Workflow Execution Service
  version: "1.0.0"
  description: >
    Runtime control plane for the BPMN-driven Workflow Engine. Manages workflow instances,
    task projections for dashboard queries, and user signals (claim, complete, defer, reassign).
    All identity and tenant context is injected by the Envoy API Gateway — callers never set
    these headers directly.
servers:
  - url: https://api.internal/execution
    description: Internal API gateway base (service prefix only; paths include /api/v1/)
security:
  - GatewayTenantContext: []
    UserContext: []
...
components:
  securitySchemes:
    GatewayTenantContext:
      type: apiKey
      in: header
      name: x-tenant-id
      description: Tenant UUID injected by the API Gateway.
    UserContext:
      type: apiKey
      in: header
      name: x-user-id
      description: Authenticated user UUID injected by the API Gateway.
    AdminRole:
      type: apiKey
      in: header
      name: x-tenant-roles
      description: >
        Comma-separated tenant roles injected by the API Gateway. Admin-only endpoints
        require at least one of: tenant_admin, tenant_owner.
    InternalServiceAuth:
      type: apiKey
      in: header
      name: x-internal-token
      description: >
        Service-to-service credential for the internal route group (RequireInternalToken,
        INTERNAL_API_TOKEN) — the same convention Definition Service already ships. These
        routes are NOT exposed through the Envoy public gateway and do NOT carry gateway
        identity headers; they are restricted to in-mesh callers by a Kubernetes
        NetworkPolicy AND authenticated by this shared secret — defense-in-depth, not either
        control alone (§9.2). The tenant context for these routes is taken from the request
        payload, not from x-tenant-id.
  parameters:
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID (injected by gateway).
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Authenticated user UUID (injected by gateway).
    TenantRolesHeader:
      name: x-tenant-roles
      in: header
      required: true
      schema:
        type: string
      description: >
        Comma-separated tenant roles. Must include tenant_admin or tenant_owner
        for admin-only endpoints.
    DepartmentsHeader:
      name: x-departments
      in: header
      required: false
      schema:
        type: string
      description: Comma-separated department UUIDs the caller belongs to (gateway-injected).
...
    InstanceIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workflow instance UUID.
...
    CursorQuery:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: >
        Opaque keyset cursor encoding the previous page's last (created_at|occurred_at, id)
        pair (§5.9). Omit for the first page. An unparseable or tampered cursor returns 400
        BAD_REQUEST rather than silently restarting.
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: >
        Page size. Values above 100 are clamped, not rejected; values <= 0 are rejected
        (§5.9).
    TraceParentHeader:
      name: traceparent
      in: header
      required: false
      schema:
        type: string
      description: >
        W3C Trace Context header (format: 00-{trace-id}-{span-id}-{flags}).
        Injected by the gateway or client to propagate distributed traces.
        Used by TracingMiddleware to link the incoming span to the existing trace.
  schemas:
...
    WorkflowEvent:
      type: object
      required: [id, workflow_instance_id, tenant_id, event_type, occurred_at]
      properties:
        id:
          type: string
          format: uuid
        workflow_instance_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
          nullable: true
          description: Set for task-scoped events; null for instance-level events.
        tenant_id:
          type: string
          format: uuid
        event_type:
          $ref: "#/components/schemas/WorkflowEventType"
        actor_user_id:
          type: string
          format: uuid
          nullable: true
          description: The user who triggered the event. Null for system/SQS-initiated transitions.
        node_key:
          type: string
          nullable: true
          description: BPMN node key for task-scoped events.
          example: "review_stage_1"
        payload_json:
          type: object
          additionalProperties: true
          description: Event-specific context (e.g. reason, new_assignee_user_id).
        occurred_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Malformed JSON body or missing required fields.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://errors.bcbp.io/execution/bad-request"
            title: "Bad Request"
            status: 400
            detail: "field 'version' is required"
            instance: "/api/v1/tasks/3fa85f64-5717-4562-b3fc-2c963f66afa6/claim"
            code: "BAD_REQUEST"
...
paths:
...
  /api/v1/instances/{id}/events:
    get:
      operationId: listInstanceEvents
      tags: [Instances]
      summary: Get activity event log for a workflow instance
      description: >
        Returns the activity event log for a workflow instance ordered by occurred_at DESC,
        indexed by (workflow_instance_id, occurred_at DESC, id DESC) (§4.9). Keyset pagination
        (§5.9), same contract as every other list endpoint — cursor/limit in, items[] +
        next_cursor out. Same intra-tenant visibility rule as GET /instances/:id (§9.2): visible
        only if the caller is (or was) an assignee on any task under this instance, the caller's
        x-departments includes any currently-open task's department_id, or the caller holds
        tenant_admin/tenant_owner.
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantRolesHeader"
        - $ref: "#/components/parameters/DepartmentsHeader"
        - $ref: "#/components/parameters/TraceParentHeader"
        - $ref: "#/components/parameters/InstanceIdPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Keyset page of the instance's activity event log (§5.9).
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkflowEvent"
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor for the next (older) page; omitted/null on the final page.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: >
            NOT_AUTHORIZED_FOR_RESOURCE — caller is not an assignee, not in the current
            department, and not an admin.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/not-authorized-for-resource"
                title: "Not Authorized For Resource"
                status: 403
                detail: "caller is not an assignee, department member, or admin for this instance"
                instance: "/api/v1/instances/3fa85f64-5717-4562-b3fc-2c963f66afa6/events"
                code: "NOT_AUTHORIZED_FOR_RESOURCE"
        "404":
          $ref: "#/components/responses/InstanceNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/internal/events:
    post:
      operationId: ingestInternalEvent
      tags: [Internal (Events)]
      summary: Ingest a forwarded domain-event envelope
      description: >
        Service-to-service ingest endpoint called by the shared Workflow-Events Consumer,
        which subscribes Execution's inbound SNS topics (iam.membership.events,
        iam.user.events, wf.template.events, §6.1) and forwards each envelope here over
        HTTP, exactly as it does for Definition Service. Sets the RLS GUC from the
        envelope's tenant_id and dispatches by type (§6.2). Idempotent: a repeated envelope
        id (paired with a fixed per-subscription consumer name) returns 2xx with no side
        effects. An unrecognized type is accepted and ignored with 200, never an error. Not
        exposed on the public Envoy gateway; carries no gateway identity headers.
      security:
        - InternalServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventEnvelope" }
      responses:
        "200":
          description: >
            Event processed, or an idempotent no-op for an already-seen envelope id, or an
            unrecognized type accepted and ignored. No response body in any case.
        "400":
          description: Malformed envelope or payload — non-retryable, the consumer routes to its DLQ.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          description: >
            Transient processing failure. The consumer retries until the queue's
            maxReceiveCount (5) moves the message to its DLQ.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "502":
          description: >
            EVENT_DECODE_FAILED — the envelope carries a dataschema (schema-registry payload)
            that could not be decoded: a missing/misconfigured decoder, or a schema-registry
            decode failure. Retryable infra problem, never a malformed client payload — the
            consumer must retry this rather than route it to its DLQ.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://errors.bcbp.io/execution/event-decode-failed"
                title: "Event Decode Failed"
                status: 502
                detail: "schema-registry payload decode failed"
                instance: "/api/v1/internal/events"
                code: "EVENT_DECODE_FAILED"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # internal, service-to-service caller — type-scoped subpaths (rev 1.36).
  # The shared Workflow-Events Consumer posts execution-bound events directly
  # to whichever of these 3 matches (internal/forwarder/category.go);
  # /api/v1/internal/events above stays registered as the catch-all for
  # workflow.task.created and any future/unrecognized type. Same auth,
  # envelope, and retry contract as the catch-all — the subpath is routing
  # sugar, not a distinct contract, so responses/security are identical.
...